LivePulse/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
tweepy==4.18.0
praw==8.5.0
transformers==4.34.0
torch==2.1.3
redis==4.8.0
python-dotenv==1.0.1
